Merge Backend to Main

+++ backend/src/main/java/com/sol/snappick/cart/service/CartService.java
@@ -4,19 +4,28 @@ import com.sol.snappick.cart.dto.CartItemReq;
 import com.sol.snappick.cart.dto.CartItemRes;
 import com.sol.snappick.cart.exception.CartItemNotFoundException;
 import com.sol.snappick.cart.exception.CartNotFoundException;
+import com.sol.snappick.member.entity.Member;
+import com.sol.snappick.member.exception.AccessDeniedException;
+import com.sol.snappick.member.repository.MemberRepository;
+import com.sol.snappick.product.dto.CartCreateRes;
 import com.sol.snappick.product.entity.Cart;
 import com.sol.snappick.product.entity.CartItem;
+import com.sol.snappick.product.entity.CartStatus;
 import com.sol.snappick.product.entity.Product;
 import com.sol.snappick.product.exception.ProductNotFoundException;
 import com.sol.snappick.product.exception.QuantityException;
 import com.sol.snappick.product.repository.CartItemRepository;
 import com.sol.snappick.product.repository.CartRepository;
 import com.sol.snappick.product.repository.ProductRepository;
+import com.sol.snappick.store.entity.Store;
+import com.sol.snappick.store.exception.StoreNotFoundException;
+import com.sol.snappick.store.repository.StoreRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.sol.snappick.product.mapper.ProductMapper;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -27,36 +36,99 @@ public class CartService {
     private final CartRepository cartRepository;
     private final ProductRepository productRepository;
     private final CartItemRepository cartItemRepository;
+    private final StoreRepository storeRepository;
+    private final MemberRepository memberRepository;
 
     private final ProductMapper productMapper;
 
     @Transactional
-    public CartItemRes createCartItem(Integer cartId, CartItemReq cartItemReq) {
+    public CartCreateRes createCart(Integer memberId, Integer storeId) throws Exception {
 
-        //cart가 존재하는지 확인한다.
+        //유효성 검증
+        //1) member가 존재하는지 확인한다.
+        Member customer = memberRepository.findById(memberId)
+                .orElseThrow(() -> new AccessDeniedException("존재하지 않는 회원입니다."));
+
+        //2) store가 존재하는지 확인한다.
+        Store store = storeRepository.findById(storeId)
+                .orElseThrow(() -> new StoreNotFoundException());
+
+        //3) CartStatus."결제대기" 중인 카트가 있는지 확인한다.
+        // 없다면 카트를 생성한다.
+        Cart cartToCreate = cartRepository.findByStoreIdAndCustomerIdAndStatus(storeId, memberId, CartStatus.결제대기);
+        if (cartToCreate==null) {
+            cartToCreate = Cart.builder()
+                    .store(store)
+                    .customer(customer)
+                    .status(CartStatus.결제대기)
+                    .items(new ArrayList<>())
+                    .build();
+        }
+        //있다면 담겨있던 cartItem을 모두 삭제한다.
+        else {
+            cartItemRepository.deleteByCartId(cartToCreate.getId());
+            cartToCreate.getItems().clear();
+        }
+        cartToCreate = cartRepository.save(cartToCreate);
+
+        //Cart: store, customer, transaction, items, status
+        //To
+        //CartRes: id, store_id, customer_id, transaction_id, items.size(), status
+
+        return CartCreateRes.builder()
+                .cartId(cartToCreate.getId())
+                .build();
+    }
+
+    @Transactional
+    public CartItemRes createCartItem(Integer memberId, Integer cartId, CartItemReq cartItemReq) throws Exception {
+
+        //유효성 검증
+        //1) cart가 존재하는지 확인한다.
         Cart cart = cartRepository.findById(cartId)
                 .orElseThrow(() -> new CartNotFoundException());
 
-        //product가 존재하는지 확인한다.
+        //2) cart 접근 권한을 확인한다.
+//        if (cart.getCustomer().getId()!=memberId)
+//            throw new AccessDeniedException();
+
+        //3) 구매하려는 product가 존재하는지 확인한다.
         Product product = productRepository.findById(cartItemReq.getProductId())
                 .orElseThrow(()->new ProductNotFoundException());
 
-        //주문 수량이 재고의 개수를 넘지 않는지 확인한다.
-        if (cartItemReq.getQuantity()>product.getStock()){
+        //4) 장바구니에 동일 상품이 있는지 확인한다.
+        CartItem cartItemToCreate = cartItemRepository.findByIdAndProductId(cartId, product.getId());
+        //없으면 새로 생성한다.
+        if (cartItemToCreate==null){
+            cartItemToCreate = CartItem.builder()
+                    .cart(cart)
+                    .product(product)
+                    .quantity(cartItemReq.getQuantity())
+                    .build();
+        }
+        //이미 존재한다면 quantity만 수정한다.
+        else {
+            cartItemToCreate.setQuantity(cartItemReq.getQuantity()+cartItemToCreate.getQuantity());
+        }
+
+        //5) 주문 수량이 재고의 개수보다 작은지 확인한다.
+        if (cartItemToCreate.getQuantity()>product.getStock()){
             throw new QuantityException();
         }
-        //주문 수량이 인당 개수 제한을 만족하는지 확인한다.
-        if (cartItemReq.getQuantity()>product.getPersonalLimit()){
+
+        //6) 주문 수량이 인당 개수 제한을 만족하는지 확인한다.
+        if (cartItemToCreate.getQuantity()>product.getPersonalLimit()){
             throw new QuantityException("인당 구매 가능 수량을 넘어서는 수량은 구매할 수 없습니다.");
         }
 
+        //TODO 7) 주문 수량이 일일 개수 제한을 만족하는지 확인한다.
+
         //cart에 cartItem을 추가한다.
-        CartItem cartItemToCreate = CartItem.builder()
-                .cart(cart)
-                .product(product)
-                .quantity(cartItemReq.getQuantity())
-                .build();
         cartItemToCreate = cartItemRepository.save(cartItemToCreate);
+
+        //CartItem: id, cart, product, quantity
+        //To
+        //CartItemRes: id, product, quantity
 
         return CartItemRes.builder()
                 .id(cartItemToCreate.getId())
@@ -66,29 +138,38 @@ public class CartService {
     }
 
     @Transactional
-    public CartItemRes updateCartItem(Integer cartId, Integer itemId, CartItemReq cartItemReq) {
+    public CartItemRes updateCartItem(Integer memberId, Integer cartId, Integer itemId, CartItemReq cartItemReq) throws Exception {
 
-        //cart가 존재하는지 확인한다.
+        //유효성 검증
+        //1) cart가 존재하는지 확인한다.
         Cart cart = cartRepository.findById(cartId)
                 .orElseThrow(() -> new CartNotFoundException());
 
-        //cartItem가 존재하는지 확인한다.
+        //2) cart 접근 권한을 확인한다.
+//        if (cart.getCustomer().getId()!=memberId)
+//            throw new AccessDeniedException();
+
+        //3) cartItem이 존재하는지 확인한다.
         CartItem cartItemToUpdate = cartItemRepository.findById(itemId)
                 .orElseThrow(() -> new CartItemNotFoundException());
 
+        //4) 주문 수량이 재고의 개수를 넘지 않는지 확인한다.
         Product product = cartItemToUpdate.getProduct();
-        //주문 수량이 재고의 개수를 넘지 않는지 확인한다.
         if (cartItemReq.getQuantity()>product.getStock()){
             throw new QuantityException();
         }
-        //주문 수량이 인당 개수 제한을 만족하는지 확인한다.
+
+        //5) 주문 수량이 인당 개수 제한을 만족하는지 확인한다.
         if (cartItemReq.getQuantity()>product.getPersonalLimit()){
             throw new QuantityException("인당 구매 가능 수량을 넘어서는 수량은 구매할 수 없습니다.");
         }
 
+        //TODO 6) 주문 수량이 일일 개수 제한을 만족하는지 확인한다.
+
         //quantity를 갱신한다.
         cartItemToUpdate.updateDetails(cartItemReq.getQuantity());
         cartItemToUpdate = cartItemRepository.save(cartItemToUpdate);
+
         return CartItemRes.builder()
                 .id(cartItemToUpdate.getId())
                 .product(productMapper.toSimpleDto(cartItemToUpdate.getProduct()))
@@ -99,9 +180,14 @@ public class CartService {
     @Transactional
     public List<CartItemRes> readCartItem(Integer cartId) {
 
-        //cart가 존재하는지 확인한다.
+        //유효성 검증
+        //1) cart가 존재하는지 확인한다.
         Cart cart = cartRepository.findById(cartId)
                 .orElseThrow(() -> new CartNotFoundException());
+
+        //2) cart 접근 권한을 확인한다.
+//        if (cart.getCustomer().getId()!=memberId)
+//            throw new AccessDeniedException();
 
         //cart에 속하는 cartItem을 조회한다.
         List<CartItem> items = cartItemRepository.findByCart(cart);
@@ -115,16 +201,21 @@ public class CartService {
     }
 
     @Transactional
-    public Boolean deleteCartItem(Integer cartId, Integer itemId) {
-        //cart가 존재하는지 확인한다.
+    public Boolean deleteCartItem(Integer memberId, Integer cartId, Integer itemId) throws Exception {
+        //유효성 검증
+        //1) cart가 존재하는지 확인한다.
         Cart cart = cartRepository.findById(cartId)
                 .orElseThrow(() -> new CartNotFoundException());
 
-        //cartItem이 존재하는지 확인한다.
+        //2) cart 접근 권한을 확인한다.
+        if (cart.getCustomer().getId()!=memberId)
+            throw new AccessDeniedException();
+
+        //3) cartItem이 존재하는지 확인한다.
         CartItem cartItemToDelete = cartItemRepository.findById(itemId)
                 .orElseThrow(() -> new CartItemNotFoundException());
 
-        //cartItem 삭제 시도한다.
+        //cartItem을 삭제한다.
         try {
             cartItemRepository.delete(cartItemToDelete);
             return true;
